Add authorisation to listings_controller.
Only authenticated user can add,edit and destroy a listing.
At the same time, when we are doing any of the actions above, we need to make sure that
the listing involved belongs to the current logged in user. Hence the need for `current_user.listings.*`.

When a listing is created, it will belong to the current logged in user.

Hide links to add/edit/destroy in views too.